+++ TDEngine2/include/ecs/CBaseSystem.h
@@ -30,14 +30,25 @@ namespace TDEngine2
 				A user can place initialization logic within this method
 			*/
 
-			TDE2_API void OnInit() override {}
+			TDE2_API void OnInit() override;
 
 			/*!
 				\brief The method is invoked when the system is destroyed
 			*/
 
-			TDE2_API void OnDestroy() override {}
+			TDE2_API void OnDestroy() override;
+
+			TDE2_API void OnActivated() override;			
+			TDE2_API void OnDeactivated() override;
+
+			/*!
+				\brief The method returns true if the system is registered and active
+			*/
+
+			TDE2_API bool IsActive() const override;
 		protected:
 			DECLARE_INTERFACE_PROTECTED_MEMBERS(CBaseSystem)
+		private:
+			bool mIsActive;
 	};
 }