+++ TDEngine2/source/ecs/CBaseSystem.cpp
@@ -1,0 +1,28 @@
+#include "../../include/ecs/CBaseSystem.h"
+
+
+namespace TDEngine2
+{
+	void CBaseSystem::OnInit()
+	{
+	}
+
+	void CBaseSystem::OnDestroy()
+	{
+	}
+
+	void CBaseSystem::OnActivated()
+	{
+		mIsActive = true;
+	}
+
+	void CBaseSystem::OnDeactivated()
+	{
+		mIsActive = false;
+	}
+
+	bool CBaseSystem::IsActive() const
+	{
+		return mIsActive;
+	}
+}